refactor duck
